Moved check_if_version_newer() to app_generic_functions.py Moved CreateRefinedVersion() to app_generic_classes.py Created app_generic_disk.py and moved get_file_content() and write_file_to_disk() to it Refactored
