+++ extras/deb_package_files/DEBIAN/control
@@ -1,5 +1,5 @@
 Package: kootnet-sensors
-Version: 0.35-92
+Version: 0.35-102
 Section: utils
 Priority: extra
 Architecture: all
