+++ http_server/flask_blueprints/atpro/remote_management/rm_receive_configs.py
@@ -19,7 +19,7 @@
 from operations_modules import logger
 from operations_modules import file_locations
 from operations_modules import app_cached_variables
-from operations_modules.app_generic_functions import write_file_to_disk
+from operations_modules.app_generic_disk import write_file_to_disk
 from configuration_modules import app_config_access
 from configuration_modules.config_primary import CreatePrimaryConfiguration
 from configuration_modules.config_urls import CreateURLConfiguration
