+++ http_server/flask_blueprints/html_functional.py
@@ -25,7 +25,8 @@ from flask import Blueprint, render_template, session, redirect, request, send_f
 from operations_modules import logger
 from operations_modules import file_locations
 from operations_modules import app_cached_variables
-from operations_modules.app_generic_functions import verify_password_to_hash, get_file_content
+from operations_modules.app_generic_functions import verify_password_to_hash
+from operations_modules.app_generic_disk import get_file_content
 from http_server.server_http_auth import auth
 from http_server.flask_blueprints.atpro.atpro_generic import get_message_page
