+++ operations_modules/initialization_checks.py
@@ -27,7 +27,8 @@ from operations_modules import logger
 from operations_modules import file_locations
 from operations_modules import software_version
 from operations_modules import app_cached_variables
-from operations_modules.app_generic_functions import write_file_to_disk, thread_function, get_file_content
+from operations_modules.app_generic_functions import thread_function
+from operations_modules.app_generic_disk import get_file_content, write_file_to_disk
 from operations_modules.sqlite_database import check_main_database_structure, check_checkin_database_structure, \
     run_database_integrity_check, check_mqtt_subscriber_database_structure
 from upgrade_modules.program_upgrade_checks import run_configuration_upgrade_checks
