+++ operations_modules/software_version.py
@@ -18,41 +18,7 @@
 """
 
 from os import path
-from operations_modules import logger
 from operations_modules import file_locations
-
-
-class CreateRefinedVersion:
-    """ Takes the provided program version as text and creates a data class object. """
-    def __init__(self, version_text=""):
-        self.major_version = 0
-        self.feature_version = 0
-        self.minor_version = 0
-        self.load_from_string(version_text)
-
-    def load_from_string(self, version_text):
-        try:
-            if len(version_text) < 15:
-                version_split = str(version_text).strip().split(".")
-                if len(version_split) == 3:
-                    self.major_version = self._convert_to_int(version_split[0])
-                    self.feature_version = self._convert_to_int(version_split[1])
-                    self.minor_version = self._convert_to_int(version_split[2])
-                else:
-                    logger.primary_logger.debug("Software Version - Invalid version text")
-        except Exception as error:
-            logger.primary_logger.debug("Software Version - Error converting text to version: " + str(error))
-
-    def get_version_string(self):
-        return str(self.major_version) + "." + str(self.feature_version) + "." + str(self.minor_version)
-
-    @staticmethod
-    def _convert_to_int(text_number):
-        try:
-            return int(text_number)
-        except Exception as error:
-            logger.primary_logger.debug("Software Version - Refined Conversion Error: " + str(error))
-            return 0
 
 
 def _get_old_version():
@@ -73,5 +39,5 @@ def write_program_version_to_file():
 
 
 # Current Version of the program
-version = "Beta.35.92"
+version = "Beta.35.102"
 old_version = _get_old_version()
